Made implementation of glCheck macro more robust
At the moment, glCheck(...) looks like a function call, which is misleading and can cause subtle bugs, for example when used in if statements. This change mitigates the situation by allowing glCheck(...) expressions to be used as single statements within if/else branches.

Initializations of OpenGL handles that previously involved glCheck() calls now need to be split into separate declaration and assignment.